+++ react-chat/index.js
@@ -1,11 +1,14 @@
 import React, {StrictMode} from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './src/index.css';
 import App from "./App";
+import AuthProvider from "./context/AuthContext";
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
     <StrictMode>
-        <App />
+        <AuthProvider>
+            <App />
+        </AuthProvider>
     </StrictMode>
 );
